mail_notifier/requirements.txt
schedule
requests
python-dotenv